requirements.txt
fastapi
uvicorn
sentence-transformers
torch
pandas
scikit-learn
ftfy
numpy
faiss-cpu
tiktoken
protobuf